Updated for fs2 3.x socket approach
